modify
git-svn-id: https://svn.baidu.com/myspace/sep/chenghuige/melt-train@276545 c4ccc3fd-b13d-49bf-a705-f618d8f4424a

+++ include/Trainers/FastRank/BinaryClassificationFastRank.h
@@ -60,7 +60,7 @@ public:
 			>> to_vector();*/
 
 		TrainSetLabels = from(TrainSet.Ratings)
-			>> select([this](short a) { return (bool)(a >= ((BinaryClassificationFastRankArguments*)(_args.get()))->smallestPositive); })
+			>> select([this](Float a) { return (bool)(a >= ((BinaryClassificationFastRankArguments*)(_args.get()))->smallestPositive); })
 			>> to_vector();
 	}
